services/api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.28
asyncpg>=0.29.0
redis>=5.0.3
boto3>=1.34.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-exporter-otlp-proto-grpc>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0
civicproof-common @ file://../packages/common
